requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
prometheus-client==0.19.0
jinja2==3.1.2
aiofiles==23.2.1
